[#162] Refactor and optimize component rendering [#162] Split ai-chat store into multiple stores [#162] Add global log() function in /utils
