+++ src-app/biz-modules/ai-chat/state/_layout-store/provider.tsx
@@ -1,0 +1,10 @@
+import { useMemo, useState } from 'react'
+import { AiChatView, LayoutContext, Store } from './context'
+
+export const LayoutProvider = ({ children }: ReactProps) => {
+	const [activeView, setActiveView] = useState<AiChatView>(AiChatView.NONE)
+
+	const store: Store = useMemo(() => ({ activeView, setActiveView }), [activeView])
+
+	return <LayoutContext.Provider value={store}>{children}</LayoutContext.Provider>
+}
