+++ src-app/biz-modules/ai-chat/state/_multi-store/_stores/_utils.ts
@@ -1,5 +1,5 @@
 import { MessageRole } from '@api/types'
-import { Message } from '../../api'
+import { Message } from '../../../api'
 
 export const GHOST_CHAT = {
 	id: 0,
