+++ src-app/biz-modules/ai-chat/state/_provider.tsx
@@ -1,38 +1,13 @@
-import { useMemo, useState } from 'react'
-import { AiChatView, Context, Store } from './_context'
-import { useAllChatsStore } from './stores/all-chats-store'
-import { useAllSubchatsStore } from './stores/all-subchats-store'
-import { useChatStore } from './stores/chat-store'
-import { useSearchStore } from './stores/search-store'
-import { useSubchatStore } from './stores/subchat-store'
+import { LayoutProvider } from './_layout-store/provider'
+import { MultiStoreProvider } from './_multi-store/provider'
+import { SearchProvider } from './_search-store/provider'
 
 export const AiChatProvider = ({ children }: ReactProps) => {
-	const [activeView, setActiveView] = useState<AiChatView>(AiChatView.NONE)
-	const allChatsStore = useAllChatsStore()
-	const chatStore = useChatStore(allChatsStore)
-	const allSubchatsStore = useAllSubchatsStore(chatStore)
-	const subchatStore = useSubchatStore(chatStore, allSubchatsStore)
-	const searchStore = useSearchStore()
-
-	const store: Store = useMemo(
-		() => ({
-			...allChatsStore,
-			...allSubchatsStore,
-			...chatStore,
-			...subchatStore,
-			...searchStore,
-			activeView,
-			setActiveView,
-		}),
-		[
-			...Object.values(allChatsStore),
-			...Object.values(chatStore),
-			...Object.values(allSubchatsStore),
-			...Object.values(subchatStore),
-			...Object.values(searchStore),
-			activeView,
-		]
+	return (
+		<LayoutProvider>
+			<SearchProvider>
+				<MultiStoreProvider>{children}</MultiStoreProvider>
+			</SearchProvider>
+		</LayoutProvider>
 	)
-
-	return <Context.Provider value={store}>{children}</Context.Provider>
 }
