+++ src-app/biz-modules/ai-chat/state/_search-store/context.ts
@@ -1,0 +1,32 @@
+import { createContext } from 'react'
+import { Chat, Message, Subchat } from '../../api'
+
+export interface SearchResult {
+	chat?: Chat
+	subchat?: Subchat
+	message?: Message
+}
+
+export interface Store {
+	showsSearch: boolean
+	searchKeyword: string
+	searchResults: SearchResult[]
+	searchPagination: Pagination
+	searchLoading: ListLoading
+	canLoadSearchResults: boolean
+	loadMoreSearchResults(): void
+	searchByKeyword(keyword: string): Promise<void>
+	setShowsSearch(value: boolean): void
+}
+
+export const SearchContext = createContext<Store>({
+	showsSearch: false,
+	searchKeyword: '',
+	searchResults: [],
+	searchPagination: { page: 0, count: 0 },
+	searchLoading: false,
+	canLoadSearchResults: false,
+	loadMoreSearchResults: () => {},
+	searchByKeyword: async () => {},
+	setShowsSearch: () => {},
+})
