+++ src-utils/@types/window.d.ts
@@ -7,6 +7,7 @@ export {}
 declare global {
 	var LOG: (...args: unknown[]) => void
 	var LOG_DEV: (...args: unknown[]) => void
+	var log: (...args: unknown[]) => void
 	var cx: (...args: ClassValue[]) => string
 	var wait: (time: number) => Promise<void>
 }
